tab bar view is working, buyers can also write reviews which are being sent to the database (still need to filter for specific suppliers and show stars rating) and can be seen under "reviews", added stem length slider to add stock and removed search bar in buyer home

+++ lib/models/stock.dart
@@ -6,7 +6,7 @@ class Stock {
   final int flowerColour;
   final String dateAdded;
   final String companyName;
-  final int stemLength;
+  final double stemLength;
 
   Stock({
     this.uid,
